Builder page: unit tests separated in different files
